@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { AppComponent } from './app.component';
@@ -10,11 +10,11 @@ import { DummyJsonService } from './dummy-service';
   declarations: [AppComponent],
   imports: [
     BrowserModule,
-    HttpClientModule,
     ReactiveFormsModule,
   ],
   providers: [
     DummyJsonService,
+    provideHttpClient(withInterceptorsFromDi()),
   ],
   bootstrap: [
     AppComponent,
